[refactor] web 디렉토리 생성 및 controller 디렉토리 위치 이동
